Move attributes to won directory and readme
